WIP raid event

+++ Scripts/PlayerData/Interface/IRaidEvent.cs
@@ -3,6 +3,6 @@
     string Id { get; set; }
     string DataId { get; set; }
     int RemainingHp { get; set; }
-    int StartTime { get; set; }
-    int EndTime { get; set; }
+    long StartTime { get; set; }
+    long EndTime { get; set; }
 }
